refactor/starting over
started testing and realized I needed a function to initialize a conversation. quickly realized that I made a lot of bad decisions along the way. conversation IDs can collide with other chains. I also want parties and chainIDs to be adjustable, which is impossible with the current context. Passing structs is bad, abi.encodePacked is a better idea. Might as well rebuild structs on the receiving end based on the decoded data. Would enable easier code flexibility. starting with a new system in which conversation IDs will be unique, but unchangeable. this will easily enable changing participants and chainIDs, allow multiple concurrent conversations between parties, and allow conversations to span multiple chains.