refactoring and additional tests for ~100% test coverage
implemented many more tests to increase coverage, and also introduced a HyperchatWithInternalFunctions.sol test contract that exposes internal function interfaces so they could be tested as well. there are 2 'branches' left to be tested, but I'm unsure of how to get them to trigger as the lines they seem to effect are already having all cases triggered. need to figure out how to map 'branches'...